refactor: resolves norbert's concerns
