--- frontend/app/components/OgImage/useOgImage.vue
@@ -1,9 +1,0 @@
-<script>
-function useOgImage() {
-  defineOgImageComponent('DefaultOg', {
-    colorMode: 'dark',
-    icon: '/img/logo.svg',
-  })
-}
-export default useOgImage
-</script>
