+++ frontend/app/utils/useOgImage.ts
@@ -1,0 +1,4 @@
+function useOgImage() {
+  defineOgImageComponent('DefaultOgImage')
+}
+export default useOgImage
